+++ src/features/migration/screens/MigrationExecute.tsx
@@ -17,6 +17,7 @@ import { DEFAULT_FULL_FAB_HEIGHT } from '@/base/components/buttons/StyledFab.tsx
 import { MigrationContinueButton } from '@/features/migration/components/MigrationContinueButton.tsx';
 import { plural } from '@lingui/core/macro';
 import { MigrationEntryGroup } from '@/features/migration/components/MIgrationEntryGroup.tsx';
+import { MigrationEntries } from '@/features/migration/MigrationEntries.ts';
 
 export const MigrationExecute = () => {
     const { t } = useLingui();
@@ -42,38 +43,29 @@ export const MigrationExecute = () => {
     const entryList = useMemo(() => Object.values(entries), [entries]);
     const migratingEntries = useMemo(
         () =>
-            entryList
-                .filter((entry) =>
-                    [MigrationEntryStatus.SEARCH_COMPLETE, MigrationEntryStatus.MIGRATING].includes(entry.status),
-                )
-                .toSorted((a, b) => {
-                    if (a.status === MigrationEntryStatus.MIGRATING && b.status !== MigrationEntryStatus.MIGRATING) {
-                        return -1;
-                    }
-
-                    if (a.status !== MigrationEntryStatus.MIGRATING && b.status === MigrationEntryStatus.MIGRATING) {
-                        return 1;
-                    }
-
-                    return entryList.indexOf(a) - entryList.indexOf(b);
-                }),
+            MigrationEntries.getActiveEntriesSorted(
+                entryList,
+                MigrationEntryStatus.MIGRATING,
+                MigrationEntryStatus.SEARCH_COMPLETE,
+                MigrationEntryStatus.MIGRATING,
+            ),
         [entryList],
     );
     const migratedEntries = useMemo(
-        () => entryList.filter((entry) => entry.status === MigrationEntryStatus.MIGRATION_COMPLETE),
+        () => MigrationEntries.getHaveStatusSorted(entryList, MigrationEntryStatus.MIGRATION_COMPLETE),
         [entryList],
     );
     const failedEntries = useMemo(
-        () => entryList.filter((entry) => entry.status === MigrationEntryStatus.MIGRATION_FAILED),
+        () => MigrationEntries.getHaveStatusSorted(entryList, MigrationEntryStatus.MIGRATION_FAILED),
         [entryList],
     );
-    const excludedEntries = useMemo(() => entryList.filter((entry) => entry.isExcluded), [entryList]);
+    const excludedEntries = useMemo(() => MigrationEntries.getExcluded(entryList), [entryList]);
     const noMatchEntries = useMemo(
-        () => entryList.filter((entry) => entry.status === MigrationEntryStatus.NO_MATCH),
+        () => MigrationEntries.getHaveStatusSorted(entryList, MigrationEntryStatus.NO_MATCH),
         [entryList],
     );
     const outdatedEntries = useMemo(
-        () => entryList.filter((entry) => entry.status === MigrationEntryStatus.OUTDATED),
+        () => MigrationEntries.getHaveStatusSorted(entryList, MigrationEntryStatus.OUTDATED),
         [entryList],
     );